check if no files are found.. if none then exit

+++ retronnex.sh
@@ -25,11 +25,16 @@ if ! git annex info; then
 fi
 
 # look in repository for files matching pattern
-FILES=$(find "${GITDIR}"/  \( ! -regex '.*/\..*' \) -type f -name ${PATTERN})
+FILES=$(find "${GITDIR}"/  \( ! -regex '.*/\..*' \) -type f -name ${PATTERN} | tr -s \n)
 
-echo -e "${BLUE}Found $(echo "${FILES}" | wc -l) files matching ${PATTERN}${NC}"
-echo "${FILES}" | sed "s|${GITDIR}./||g" | nl
-echo -e "\n Removing from git and adding to annex"
+if [ ! "${FILES}" ]; then
+	echo "No Files Found with pattern ${PATTERN} in ${GITDIR}"
+	exit
+else
+	echo -e "${BLUE}Found $(echo "${FILES}" | wc -l) files matching ${PATTERN}${NC}"
+	echo "${FILES}" | sed "s|${GITDIR}./||g" | nl
+	echo -e "\n Removing from git and adding to annex"
+	# remove files from git and annex them
+	git filter-branch --tree-filter 'for '${FILES}';do if [ -f "$FILE" ] && [ ! -L "$FILE" ];then git rm --cached "$FILE";git annex add "$FILE";ln -sf `readlink "$FILE"|sed -e "s:^../../::"` "$FILE";fi;done' --tag-name-filter cat -- --all	
+fi
 
-# remove files from git and annex them
-git filter-branch --tree-filter 'for '${FILES}';do if [ -f "$FILE" ] && [ ! -L "$FILE" ];then git rm --cached "$FILE";git annex add "$FILE";ln -sf `readlink "$FILE"|sed -e "s:^../../::"` "$FILE";fi;done' --tag-name-filter cat -- --all
